+++ app/__init__.py
@@ -1,0 +1,3 @@
+from .run_pipeline import run_pipeline
+
+__all__ = ['run_pipeline']
