+++ compose.yml
@@ -1,0 +1,6 @@
+services:
+  data-pipeline:
+    build: .
+    container_name: simple_pipeline_container
+    volumes:
+      - ./data:/data
