+++ main.py
@@ -1,0 +1,7 @@
+from app import run_pipeline
+
+def main():
+    return run_pipeline()
+    
+if __name__ == "__main__":
+    main()
