+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "simple-pipeline"
-version = "0.1.0"
-description = "Add your description here"
+version = "1.1.0"
+description = "Pipeline to process csv file data and deliver it into another csv file"
 readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
